Refactor SenderProcessor and allow it's reuse in connectors
